Moved interface file
